+++ app/src/slices/admin/adminSlice.tsx
@@ -1,7 +1,4 @@
 import { createSlice, PayloadAction,createAsyncThunk  } from "@reduxjs/toolkit";
-import {LOCAL_URL} from '../../constants/Config'
-import axios from "axios";
-import jwt_decode from 'jwt-decode'
 
 
 
@@ -25,17 +22,7 @@ const initialState : AdminState ={
     
 };  
 
-export const loginAdmin = createAsyncThunk(
-'/admins/loginAdmin',async(adminCredential:AdminState)=>{
-    
-const Response = await axios.post(`${LOCAL_URL}/auth/loginAdmin`,adminCredential.admin)
-console.log(Response.data.token);
-localStorage.setItem('admin',JSON.stringify(Response))
-return Response.data;
 
-}
-
-);
 
 const adminSlice = createSlice({
     name:'admins',
@@ -43,27 +30,6 @@ const adminSlice = createSlice({
     reducers: {
       
     },
-    extraReducers:(builder)=>{
-
-        builder.addCase(loginAdmin.pending,(state)=>{
-            state.loading=true;
-            state.error=null;
-        })
-        .addCase(loginAdmin.fulfilled,(state,action)=>{
-            state.loading=false;
-            state.admin=action.payload;
-            state.error=null;
-        }) 
-        .addCase(loginAdmin.rejected,(state,action)=>{
-            state.loading=false;
-            if (action.payload && typeof action.payload == "string") {
-                state.error = action.payload;
-            }
-           
-           //state.error=action.error.message;
-        })
-    }
-
 }
 );
