--- app/src/slices/authSlice.ts
@@ -1,11 +1,0 @@
-import {createSlice} from '@reduxjs/toolkit'
-
-
-
-const authSlice = createSlice({
-
-    name:'auth',
-    initialState:'',
-    reducers:{},
-});
-export default authSlice.reducer;
