+++ app/src/store/configureStore.tsx
@@ -1,11 +1,12 @@
 import {configureStore} from "@reduxjs/toolkit"
 import adminReducer  from "../slices/admin/adminSlice"
 import professionalReducer from '../slices/professionals/ProfessionalsSlice'
-
+import adminAuthReducer from '../slices/admin/authSlice'
 
 export const store = configureStore({
 
 reducer:{
+    adminAuth : adminAuthReducer,
     admins : adminReducer, 
     professionals: professionalReducer,
